feat: Migrate to typescript
