+++ src/components/SponsorsSection.tsx
