+++ src/declarations.d.ts
@@ -1,0 +1,5 @@
+declare module '*.jpeg';
+declare module '*.jpg';
+declare module '*.png';
+declare module '*.svg';
+declare module '*.css';
